YEAR_3/Application_Development/AREA_first_version/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
django-allauth==0.57.0
python-dotenv==1.0.0
requests==2.31.0